feat: implement update and delete functionality for students; add UpdateStudentDto and enhance success messages

+++ server/src/lecturers/lecturers.service.ts
@@ -172,6 +172,6 @@ export class LecturersService {
 
     // TODO: Check if the lecturer is assigned to any courses before deleting
 
-    return await this.usersRepository.delete({ username });
+    return await this.usersRepository.delete({ username, roleId: 'lecturer' });
   }
 }
